Refactor CurrentMap and Screen out
Add a map reference to PositionComponent. It was unnatural
for a position to exist independently of the map.
Access the "current map" via the player, instead of a static global.
Replace the global Screen with a CameraComponent.

As a welcome side effect, the above changes open a possibility of
split-screen in the future. FadeSystem is not split-screen ready yet,
though.